got project build working in visual studio, refactored a bit of stuff like z index, deleted makefile

+++ include/game_state.hpp
@@ -13,7 +13,7 @@ class GameState
 public:
 
     EventBus event_bus;
-    virtual void draw(sf::RenderWindow& window, int elapsed_ms) = 0;
+    virtual void draw(sf::RenderWindow& window, double elapsed_ms) = 0;
 };
 
 #endif
