+++ include/player.hpp
@@ -13,7 +13,7 @@ public:
 
     Player(const TextureManager& texture_manager);
     void walk();
-    void draw(sf::RenderWindow& window, int elapsed_ms);
+    void draw(sf::RenderWindow& window, double elapsed_ms);
 };
 
 #endif
