+++ src/grass_tile.cpp
@@ -5,7 +5,7 @@ GrassTile::GrassTile(const TextureManager& texture_manager) : MapTile()
     this->sprite.setTexture(texture_manager.get_texture("grass_tile"));
 }
 
-void GrassTile::draw(sf::RenderWindow& window, int elapsed_ms)
+void GrassTile::draw(sf::RenderWindow& window, double elapsed_ms)
 {
     window.draw(this->sprite);
 }
