+++ src/map.cpp
@@ -16,9 +16,9 @@ Map::Map(const TextureManager& texture_manager, int num_tiles_x, int num_tiles_y
     }
 }
 
-void Map::draw(sf::RenderWindow& window, int elapsed_ms)
+void Map::draw(sf::RenderWindow& window, double elapsed_ms)
 {
-    for (int i = 0; i < tiles.size(); i ++)
+    for (size_t i = 0; i < tiles.size(); i ++)
     {
         tiles[i]->draw(window, elapsed_ms);
     }
